prototypee must not inherit from protype's class directly
Given a class, an instance of the class and an object using the
instance as prototype. The class defines a class variable, the
instance overrides it with an instance variable; If the prototypee
also inherits from the class it will find the class (default) variable
before the more specific instance variable. Therefore the prototypee
may not inherit from the prototype's class directly.